Adds a logging to build reports.
Adds logging system to build a result xml file for
each module of a maven project. For a simple project,
the result file will be generate to target/spoon-maven-plugin
folder of the project. For multi module project, the
result file will be generate in each target/spoon-maven-plugin
folders of each module.