Can launch the plugin several times without clean maven command.
Before, when the plugin identifies a result file in the
build directory (target), the plugin stopped and notified
the developer that the project is already spooned. Now,
the result file contains a timestamp in its filename
and the project can install several times a project
given without any clean command.